move compatability test to mixin from base
